Extract message processing to simplify the design (and tests)
