+++ .gitignore
@@ -14,3 +14,5 @@ test.properties
 # GitHub Actions #
 ##################
 .github/.release
+src/test/resources/features
+src/test/resources/feature-drafts
