+++ src/main/kotlin/com/pubnub/api/models/consumer/pubsub/PNEvent.kt
@@ -1,0 +1,3 @@
+package com.pubnub.api.models.consumer.pubsub
+
+interface PNEvent
