+++ src/main/kotlin/com/pubnub/api/models/consumer/pubsub/PNPresenceEventResult.kt
@@ -39,4 +39,4 @@ data class PNPresenceEventResult(
     val timeout: List<String>? = null,
     val hereNowRefresh: Boolean? = null,
     val userMetadata: Any? = null
-)
+) : PNEvent
