+++ src/main/kotlin/com/pubnub/api/models/consumer/pubsub/files/PNFileEventResult.kt
@@ -2,6 +2,7 @@ package com.pubnub.api.models.consumer.pubsub.files
 
 import com.google.gson.JsonElement
 import com.pubnub.api.models.consumer.files.PNDownloadableFile
+import com.pubnub.api.models.consumer.pubsub.PNEvent
 
 data class PNFileEventResult(
     val channel: String,
@@ -10,4 +11,4 @@ data class PNFileEventResult(
     val message: Any?,
     val file: PNDownloadableFile,
     val jsonMessage: JsonElement
-)
+) : PNEvent
