+++ src/test/kotlin/com/pubnub/contract/RunMainCucumberTest.kt
@@ -6,14 +6,16 @@ import org.junit.runner.RunWith
 
 @RunWith(Cucumber::class)
 @CucumberOptions(
-    tags = "not @skip and not @na=ruby and not @beta",
+    features = ["src/test/resources"],
+    tags = "not @skip and not @na=kotlin and not @beta and @featureSet=objectsv2",
     plugin = ["pretty", "summary", "junit:build/reports/cucumber-reports/main.xml"]
 )
 class RunMainCucumberTest
 
 @RunWith(Cucumber::class)
 @CucumberOptions(
-    tags = "not @skip and not @na=ruby and @beta",
+    features = ["src/test/resources"],
+    tags = "not @skip and not @na=kotlin and @beta and @featureSet=objectsv2",
     plugin = ["pretty", "summary", "junit:build/reports/cucumber-reports/beta.xml"]
 )
 class RunBetaCucumberTest
